delete table page. move data-grid to Dashboard.tsx, and update Dashboard.tsx
